requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
Discord-Webhooks==1.0.4
discord.py==1.7.1
idna==2.8
keyboard==0.13.5
multidict==5.1.0
numpy==1.20.2
opencv-contrib-python==4.5.1.48
pkg-resources==0.0.0
python-dotenv==0.17.0
requests==2.21.0
schedule==0.6.0
selenium==3.141.0
typing-extensions==3.7.4.3
urllib3==1.24.3
yarl==1.6.3

imutils~=0.5.4